+++ tests/test_host_vm.nim
@@ -91,7 +91,7 @@ template runTest(testName: string, create_vm, get_host_interface, create_host_co
     destination: address,
     value: balance,
     input_data: cast[ptr byte](input[0].addr),
-    input_size: input.len.uint,
+    input_size: input.len.csize_t,
     gas: gas,
     depth: 0
   )
